// uvcanvas/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    BackgroundOptionSet: ComponentFramework.PropertyTypes.EnumProperty<"Lumiflex" | "Novatrix" | "Velustro" | "Tranquiluxe" | "Opulento" | "Xenon">;
    canvasWidth: ComponentFramework.PropertyTypes.WholeNumberProperty;
    canvasHeight: ComponentFramework.PropertyTypes.WholeNumberProperty;
}
export interface IOutputs {
}